JsonVisualizer 1.0.3

+++ src/Structure.cs
@@ -6,6 +6,6 @@ namespace FreakzDEV.JsonVisualizer
     public string Text { get; set; } = string.Empty;
     public string Size { get; set; } = string.Empty;
     public string Attributes { get; set; } = string.Empty;
-    public int MarginTop { get; set; } = 0;
+    public double MarginTop { get; set; } = 0;
   }
 }
